# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Golf Charity Subscription Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Subscriptions
  /subscriptions/my:
    get:
      operationId: getMySubscription
      tags: [subscriptions]
      summary: Get current user's subscription
      responses:
        "200":
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: No active subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subscriptions:
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all subscriptions (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, cancelled, expired, pending]
      responses:
        "200":
          description: Subscriptions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
  /subscriptions/{subscriptionId}/cancel:
    post:
      operationId: cancelSubscription
      tags: [subscriptions]
      summary: Cancel a subscription
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # Scores
...
  /scores/user/{userId}:
    get:
      operationId: getUserScores
      tags: [scores]
      summary: Get a specific user's scores (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Score list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"

  # Charities
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        charityId:
          type: integer
          nullable: true
        charityPercentage:
          type: number
          nullable: true
        subscriptionStatus:
          type: string
          enum: [active, inactive, cancelled, expired]
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        charityId:
          type: integer
          nullable: true
        charityPercentage:
          type: number
          minimum: 10
          maximum: 100
    PaginatedUsers:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit

    # Subscription
    Subscription:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        plan:
          type: string
          enum: [monthly, yearly]
        status:
          type: string
          enum: [active, cancelled, expired, pending]
        amount:
          type: number
        prizeContribution:
          type: number
        charityContribution:
          type: number
        startDate:
          type: string
          format: date-time
        renewalDate:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - plan
        - status
        - amount
        - prizeContribution
        - charityContribution
        - startDate
        - createdAt
    CreateSubscriptionBody:
      type: object
      properties:
        plan:
          type: string
          enum: [monthly, yearly]
        charityId:
          type: integer
          nullable: true
        charityPercentage:
          type: number
          minimum: 10
          maximum: 100
          default: 10
      required:
        - plan

    # Score
    Score:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        value:
          type: integer
          minimum: 1
          maximum: 45
        playedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - value
        - playedAt
        - createdAt
...